Move 'timeout' to common
