+++ agent/tools/code_exec.py
@@ -22,7 +22,7 @@ from typing import Optional
 from pydantic import BaseModel, Field, field_validator
 from agent.tools.base import ToolParamBase, ToolBase, ToolMeta
 from api import settings
-from api.utils.api_utils import timeout
+from common.connection_utils import timeout
 
 
 class Language(StrEnum):
