+++ agent/tools/googlescholar.py
@@ -19,7 +19,7 @@ import time
 from abc import ABC
 from scholarly import scholarly
 from agent.tools.base import ToolMeta, ToolParamBase, ToolBase
-from api.utils.api_utils import timeout
+from common.connection_utils import timeout
 
 
 class GoogleScholarParam(ToolParamBase):
