+++ api/utils/api_utils.py
@@ -13,17 +13,14 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 #
-import asyncio
+
 import functools
 import json
 import logging
 import os
-import queue
-import threading
 import time
 from copy import deepcopy
 from functools import wraps
-from typing import Any, Callable, Coroutine, Optional, Type, Union
 
 import requests
 import trio
@@ -43,6 +40,7 @@ from api.db import ActiveEnum
 from api.db.db_models import APIToken
 from api.utils.json_encode import CustomJSONEncoder
 from rag.utils.mcp_tool_call_conn import MCPToolCallSession, close_multiple_mcp_toolcall_sessions
+from common.connection_utils import timeout
 
 requests.models.complexjson.dumps = functools.partial(json.dumps, cls=CustomJSONEncoder)
 
@@ -603,82 +601,6 @@ def get_mcp_tools(mcp_servers: list, timeout: float | int = 10) -> tuple[dict, s
     except Exception as e:
         return {}, str(e)
 
-
-TimeoutException = Union[Type[BaseException], BaseException]
-OnTimeoutCallback = Union[Callable[..., Any], Coroutine[Any, Any, Any]]
-
-
-def timeout(seconds: float | int | str = None, attempts: int = 2, *, exception: Optional[TimeoutException] = None, on_timeout: Optional[OnTimeoutCallback] = None):
-    if isinstance(seconds, str):
-        seconds = float(seconds)
-    def decorator(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            result_queue = queue.Queue(maxsize=1)
-
-            def target():
-                try:
-                    result = func(*args, **kwargs)
-                    result_queue.put(result)
-                except Exception as e:
-                    result_queue.put(e)
-
-            thread = threading.Thread(target=target)
-            thread.daemon = True
-            thread.start()
-
-            for a in range(attempts):
-                try:
-                    if os.environ.get("ENABLE_TIMEOUT_ASSERTION"):
-                        result = result_queue.get(timeout=seconds)
-                    else:
-                        result = result_queue.get()
-                    if isinstance(result, Exception):
-                        raise result
-                    return result
-                except queue.Empty:
-                    pass
-            raise TimeoutError(f"Function '{func.__name__}' timed out after {seconds} seconds and {attempts} attempts.")
-
-        @wraps(func)
-        async def async_wrapper(*args, **kwargs) -> Any:
-            if seconds is None:
-                return await func(*args, **kwargs)
-
-            for a in range(attempts):
-                try:
-                    if os.environ.get("ENABLE_TIMEOUT_ASSERTION"):
-                        with trio.fail_after(seconds):
-                            return await func(*args, **kwargs)
-                    else:
-                        return await func(*args, **kwargs)
-                except trio.TooSlowError:
-                    if a < attempts - 1:
-                        continue
-                    if on_timeout is not None:
-                        if callable(on_timeout):
-                            result = on_timeout()
-                            if isinstance(result, Coroutine):
-                                return await result
-                            return result
-                        return on_timeout
-
-                    if exception is None:
-                        raise TimeoutError(f"Operation timed out after {seconds} seconds and {attempts} attempts.")
-
-                    if isinstance(exception, BaseException):
-                        raise exception
-
-                    if isinstance(exception, type) and issubclass(exception, BaseException):
-                        raise exception(f"Operation timed out after {seconds} seconds and {attempts} attempts.")
-
-                    raise RuntimeError("Invalid exception type provided")
-
-        if asyncio.iscoroutinefunction(func):
-            return async_wrapper
-        return wrapper
-
-    return decorator
 
 
 async def is_strong_enough(chat_model, embedding_model):
